Tridm

+++ 1. Source code/Etailor.API/Etailor.API.WebAPI/Controllers/BlogController.cs
@@ -4,7 +4,9 @@ using Etailor.API.Service.Interface;
 using Etailor.API.Service.Service;
 using Etailor.API.Ultity.CustomException;
 using Etailor.API.WebAPI.ViewModels;
+using System.Security.Claims;
 using Microsoft.AspNetCore.Mvc;
+using Etailor.API.Ultity.CommonValue;
 
 namespace Etailor.API.WebAPI.Controllers
 {
@@ -13,15 +15,16 @@ namespace Etailor.API.WebAPI.Controllers
     public class BlogController : Controller
     {
         private readonly IBlogService blogService;
-        private readonly IMapper mapper;
+        private readonly IStaffService staffService;
         private readonly ICategoryService categoryService;
+        private readonly IMapper mapper;
 
-        public BlogController(IBlogService blogService, IMapper mapper, ICategoryService categoryService)
+        public BlogController(IBlogService blogService, IMapper mapper, ICategoryService categoryService, IStaffService staffService)
         {
             this.blogService = blogService;
             this.mapper = mapper;
             this.categoryService = categoryService;
-
+            this.staffService = staffService;
         }
 
         [HttpGet("{id}")]
@@ -59,7 +62,7 @@ namespace Etailor.API.WebAPI.Controllers
             try
             {
                 var blogs = blogService.GetBlogs(search);
-                return Ok(mapper.Map<IEnumerable<BlogVM>>(blogs));
+                return Json(mapper.Map<IEnumerable<ListOfBlogVM>>(blogs));
             }
             catch (UserException ex)
             {
@@ -76,39 +79,41 @@ namespace Etailor.API.WebAPI.Controllers
         }
 
         [HttpPost]
-        public IActionResult CreateBlog([FromBody] BlogVM blogVM)
+        public IActionResult CreateBlog([FromBody] CreateBlogVM blogVM)
         {
             try
             {
-                //var role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
-                //if (role == null)
-                //{
-                //    return Unauthorized("Chưa đăng nhập");
-                //}
-                //else if (role != RoleName.MANAGER)
-                //{
-                //    return Forbid("Không có quyền truy cập");
-                //}
-                //else
-                //{
-                //    var staffId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
-                //    var secrectKey = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.CookiePath)?.Value;
-                //    if (!staffService.CheckSecrectKey(staffId, secrectKey))
-                //    {
-                //        return Unauthorized("Chưa đăng nhập");
-                //    }
-                //    else
-                //    {
-                if (string.IsNullOrWhiteSpace(blogVM.Title))
+                var role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (role == null)
                 {
-                    throw new UserException("Nhập tên tiêu đề");
+                    return Unauthorized("Chưa đăng nhập");
+                }
+                else if (role != RoleName.MANAGER)
+                {
+                    return Forbid("Không có quyền truy cập");
                 }
                 else
                 {
-                    return blogService.CreateBlog(mapper.Map<Blog>(blogVM)) ? Ok("Tạo mới bài blog thành công") : BadRequest("Tạo mới bài blog thất bại");
+                    var staffId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+                    var secrectKey = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.CookiePath)?.Value;
+                    if (!staffService.CheckSecrectKey(staffId, secrectKey))
+                    {
+                        return Unauthorized("Chưa đăng nhập");
+                    }
+                    else
+                    {
+                        var blog = mapper.Map<Blog>(blogVM);
+                        if (string.IsNullOrWhiteSpace(blogVM.Title))
+                        {
+                            throw new UserException("Nhập tên tiêu đề");
+                        }
+                        else
+                        {
+                            blog.StaffId = staffId;
+                            return blogService.CreateBlog(blog) ? Ok("Tạo mới bài blog thành công") : BadRequest("Tạo mới bài blog thất bại");
+                        }
+                    }
                 }
-                //    }
-                //}
             }
             catch (UserException ex)
             {
@@ -125,46 +130,48 @@ namespace Etailor.API.WebAPI.Controllers
         }
 
         [HttpPut("{id}")]
-        public IActionResult UpdateBodySize(string? id, [FromBody] BlogVM blogVM)
+        public IActionResult UpdateBodySize(string? id, [FromBody] UpdateBlogVM blogVM)
         {
             try
             {
-                //var role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
-                //if (role == null)
-                //{
-                //    return Unauthorized("Chưa đăng nhập");
-                //}
-                //else if (role != RoleName.MANAGER)
-                //{
-                //    return Forbid("Không có quyền truy cập");
-                //}
-                //else
-                //{
-                //    var staffId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
-                //    var secrectKey = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.CookiePath)?.Value;
-                //    if (!staffService.CheckSecrectKey(staffId, secrectKey))
-                //    {
-                //        return Unauthorized("Chưa đăng nhập");
-                //    }
-                //    else
-                //    {
-                //        if (string.IsNullOrWhiteSpace(discount.Name) || string.IsNullOrEmpty(discount.Code))
-                //        {
-                //            throw new UserException("Nhập tên phiếu giảm giá");
-                //        }
-                //        else
-                //        {
-                if (string.IsNullOrEmpty(id) || string.IsNullOrEmpty(blogVM.Id) || id != blogVM.Id)
+                var role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (role == null)
                 {
-                    return NotFound("Không tìm thấy bài blog");
+                    return Unauthorized("Chưa đăng nhập");
+                }
+                else if (role != RoleName.MANAGER)
+                {
+                    return Forbid("Không có quyền truy cập");
                 }
                 else
                 {
-                    return blogService.UpdateBlog(mapper.Map<Blog>(blogVM)) ? Ok("Cập nhật bài blog thành công") : BadRequest("Cập nhật bài blog thất bại");
+                    var staffId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+                    var secrectKey = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.CookiePath)?.Value;
+                    if (!staffService.CheckSecrectKey(staffId, secrectKey))
+                    {
+                        return Unauthorized("Chưa đăng nhập");
+                    }
+                    else
+                    {
+                        var blog = mapper.Map<Blog>(blogVM);
+                        if (string.IsNullOrWhiteSpace(blog.Title))
+                        {
+                            throw new UserException("Nhập tiêu đề bài blog");
+                        }
+                        else
+                        {
+                            if (string.IsNullOrEmpty(id) || string.IsNullOrEmpty(blogVM.Id) || id != blogVM.Id)
+                            {
+                                return NotFound("Không tìm thấy bài blog");
+                            }
+                            else
+                            {
+                                blog.StaffId = staffId;
+                                return blogService.UpdateBlog(blog) ? Ok("Cập nhật bài blog thành công") : BadRequest("Cập nhật bài blog thất bại");
+                            }
+                        }
+                    }
                 }
-                //        }
-                //    }
-                //}
             }
             catch (UserException ex)
             {
@@ -185,35 +192,35 @@ namespace Etailor.API.WebAPI.Controllers
         {
             try
             {
-                //var role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
-                //if (role == null)
-                //{
-                //    return Unauthorized("Chưa đăng nhập");
-                //}
-                //else if (role != RoleName.MANAGER)
-                //{
-                //    return Forbid("Không có quyền truy cập");
-                //}
-                //else
-                //{
-                //    var staffId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
-                //    var secrectKey = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.CookiePath)?.Value;
-                //    if (!staffService.CheckSecrectKey(staffId, secrectKey))
-                //    {
-                //        return Unauthorized("Chưa đăng nhập");
-                //    }
-                //    else
-                //    {
-                if (string.IsNullOrEmpty(id))
+                var role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (role == null)
                 {
-                    return NotFound("Không tìm thấy bài blog");
+                    return Unauthorized("Chưa đăng nhập");
+                }
+                else if (role != RoleName.MANAGER)
+                {
+                    return Forbid("Không có quyền truy cập");
                 }
                 else
                 {
-                    return blogService.DeleteBlog(id) ? Ok("Xóa bài blog thành công") : BadRequest("Xóa bài blog thất bại");
+                    var staffId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+                    var secrectKey = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.CookiePath)?.Value;
+                    if (!staffService.CheckSecrectKey(staffId, secrectKey))
+                    {
+                        return Unauthorized("Chưa đăng nhập");
+                    }
+                    else
+                    {
+                        if (string.IsNullOrEmpty(id))
+                        {
+                            return NotFound("Không tìm thấy bài blog");
+                        }
+                        else
+                        {
+                            return blogService.DeleteBlog(id) ? Ok("Xóa bài blog thành công") : BadRequest("Xóa bài blog thất bại");
+                        }
+                    }
                 }
-                //    }
-                //}
             }
             catch (UserException ex)
             {
